code cleanup 2
