+++ src/lpcUtils.js
@@ -63,3 +63,98 @@ export function lpc(signal, p) {
 
     return { a, err };
 }
+
+/**
+ * Draws the spectral envelope on the canvas, scaled to 0-5000Hz.
+ * Also adds frequency labels to the X-axis.
+ * @param {Object} params The parameters for drawing.
+ * @param {Float32Array} params.lpcCoefficients The LPC coefficients.
+ * @param {number} params.sampleRate The sample rate of the audio.
+ * @param {Object} params.canvasContext The canvas context to draw on.
+ * @param {Object} params.vowelStimuli The vowel stimuli data.
+ * @param {string} params.selectedVowel The selected vowel.
+ */
+export function drawSpectralEnvelope({ lpcCoefficients, sampleRate, canvasContext, vowelStimuli, selectedVowel }) {
+    const ctx = canvasContext;
+    const canvas = ctx.canvas;
+    const sr = sampleRate;
+
+    ctx.fillStyle = '#1a202c';
+    ctx.fillRect(0, 0, canvas.width, canvas.height);
+    const maxFreq = 5000;
+    ctx.lineWidth = 1;
+    ctx.strokeStyle = '#374151';
+    ctx.fillStyle = '#9ca3af';
+    ctx.font = '12px Inter';
+
+    // Draw X-axis labels and grid lines
+    const numGridLines = 5;
+    for (let i = 0; i <= numGridLines; i++) {
+        const freq = (maxFreq / numGridLines) * i;
+        const x = (freq / maxFreq) * canvas.width;
+        if (i > 0) {
+            ctx.beginPath();
+            ctx.moveTo(x, 0);
+            ctx.lineTo(x, canvas.height - 20);
+            ctx.stroke();
+        }
+        const label = `${freq / 1000}k`;
+        ctx.fillText(label, x - (i === 0 ? 0 : 10), canvas.height - 5);
+    }
+    ctx.beginPath();
+    ctx.moveTo(0, canvas.height - 20);
+    ctx.lineTo(canvas.width, canvas.height - 20);
+    ctx.stroke();
+
+    // Calculate and Draw LPC Curve in dB
+    ctx.strokeStyle = '#4299e1';
+    ctx.lineWidth = 2;
+    ctx.beginPath();
+
+    const numPoints = canvas.width;
+    const freqResponse = new Float32Array(numPoints);
+    let maxDb = -Infinity, minDb = Infinity;
+
+    for (let i = 0; i < numPoints; i++) {
+        const freq = (i / numPoints) * maxFreq;
+        const w = 2 * Math.PI * freq / sr;
+        let re = 1.0, im = 0.0;
+        for (let k = 1; k < lpcCoefficients.length; k++) {
+            re += lpcCoefficients[k] * Math.cos(k * w);
+            im += lpcCoefficients[k] * Math.sin(k * w);
+        }
+        const mag = 1.0 / Math.sqrt(re * re + im * im);
+        const db = 20 * Math.log10(mag + 1e-12);
+        freqResponse[i] = db;
+        if (db > maxDb) maxDb = db;
+        if (db < minDb) minDb = db;
+    }
+
+    for (let i = 0; i < numPoints; i++) {
+        const x = i;
+        const y = ((maxDb - freqResponse[i]) / (maxDb - minDb)) * (canvas.height - 25) + 5;
+        if (i === 0) {
+            ctx.moveTo(x, y);
+        } else {
+            ctx.lineTo(x, y);
+        }
+    }
+    ctx.stroke();
+
+    if (!vowelStimuli || !vowelStimuli["Vowel"][selectedVowel]) {
+        console.error('Selected vowel data not found:', selectedVowel);
+        return;
+    }
+    const f1_range = vowelStimuli["Vowel"][selectedVowel]["formant"]["f1"];
+    const f2_range = vowelStimuli["Vowel"][selectedVowel]["formant"]["f2"];
+    const f1_start = (f1_range[0] / maxFreq) * canvas.width;
+    const f1_end = (f1_range[1] / maxFreq) * canvas.width;
+    const f2_start = (f2_range[0] / maxFreq) * canvas.width;
+    const f2_end = (f2_range[1] / maxFreq) * canvas.width;
+    ctx.fillStyle = 'rgba(245, 101, 101, 0.2)';
+    ctx.strokeStyle = '#f56565';
+    ctx.fillRect(f1_start, 0, f1_end - f1_start, canvas.height - 20);
+    ctx.strokeRect(f1_start, 0, f1_end - f1_start, canvas.height - 20);
+    ctx.fillRect(f2_start, 0, f2_end - f2_start, canvas.height - 20);
+    ctx.strokeRect(f2_start, 0, f2_end - f2_start, canvas.height - 20);
+}
